fix(debt): remove final integration dependency cycles
