+++ docs/source-of-truth-index.md
@@ -29,8 +29,9 @@
 - Release surface 当前真值：production builds/package outputs 不包含 dev/test harness HTML/JS；dev builds 保留 harness 页面；`audit:release-surface:report` 校验 manifest 文件引用与 forbidden harness package members，并已接入 `quality` 与 CI release-surface 步骤；Chrome ZIP 与 Firefox XPI package 脚本会先解包最终产物，再对解包目录执行同一 release-surface 审计
 - Quality gate 当前真值：`quality` 包含 i18n catalog drift check、locale source alignment、hardcoded config 守卫、type/lint ratchet、release surface 与 non-production source hard gate；CI i18n catalog drift check、locale source alignment 与 hardcoded config guard 均为 hard gate；`verify:preflight` 同样显式包含 `i18n:catalog:check`；`lint:options-css` 对当前 `src/options/stitch/styles/**` 解析出非空 `selector-class-pattern` 规则
 - 2026-06-13 P06 final ratchet current truth：`audit:build:report` 继续识别 `.generated-*` release locale chunks；generated locale modules 仍保持 non-schema runtime messages + WebExtension static messages split。当前 integration dev build exact stop gates 为 `content/runtime.js` raw `57,386` bytes、`onboarding/index.js` raw `16,459` bytes、`chunk count <= 117`；没有放宽 locale、single/shared chunk 或 YAML budgets。
-- 2026-06-13 P06 type/test/platform/non-production truth：`lint:type-any` 当前扫描 `1172` files，fresh overall `0/1121/1869/45/3`、src `0/609/630/9/0`、tests `0/512/1239/36/3`；checked-in `lint:type-any:ratchet` 上限当前为 overall `0/1125/1869/53/4`、src `0/613/630/9/0`、tests `0/513/1239/46/4`，本次只同步 assertions，`any` 继续保持 `0`，`ts-expect-error` 未放宽；`audit:platform-boundary:report` 当前仍为 `148` findings（composition-root `11`、offscreen-local-vault-permission-root `1`、platform-adapter `93`、shared-runtime-helper `23`、type-only `20`）；`audit:non-production-source:report` 在 `audit:production-build-graph:report` 之后退出 `0`，decision counts 仍为 `retain-production: 624`、`migrate-import-owner: 134`、`retain-production-facade: 17`。
+- 2026-06-13 P06 type/test/platform/non-production truth：`lint:type-any` 当前扫描 `1173` files，fresh overall `0/1121/1869/45/3`、src `0/609/630/9/0`、tests `0/512/1239/36/3`；checked-in `lint:type-any:ratchet` 上限当前为 overall `0/1125/1869/53/4`、src `0/613/630/9/0`、tests `0/513/1239/46/4`，本次只同步 assertions，`any` 继续保持 `0`，`ts-expect-error` 未放宽；`audit:platform-boundary:report` 当前仍为 `148` findings（composition-root `11`、offscreen-local-vault-permission-root `1`、platform-adapter `93`、shared-runtime-helper `23`、type-only `20`）；`audit:non-production-source:report` 在 `audit:production-build-graph:report` 之后退出 `0`，decision counts 为 `retain-production: 625`、`migrate-import-owner: 134`、`retain-production-facade: 17`。
 - 2026-06-13 P06 inherited video session test truth：当前 P06 分支的 `tests/unit/content/video/VideoSession.test.ts` 已通过 focused restored screenshot run、same-page owner-context regression 与 full-file run。P06 只在测试内等待 restored timestamp screenshot payload 写回状态，并显式接入 same-page owner-context runtime messenger；没有修改 production `src/**`。
+- 2026-06-13 final integration dependency-cycle truth：`audit:deps:report` 当前为 `modules=859`、`dependencies=2558`、`violations=0`。截图准备请求状态已拆入 `src/content/video/videoScreenshotPreparationRequestStore.ts`，coordinator 不再从 queue implementation import type；i18n pseudo locale、locale service 与 pseudo localization 均通过 `localeDefinition.ts` 取得 `RuntimeMessages` 类型，不再从 runtime `messages.ts` 回拉，production `messages.ts` bundle 也由 `tests/unit/i18n/config.test.ts` 守住不得包含 `qps-ploc` 或 pseudo marker。
 - i18n 语言范围当前真值：产品决策标签为 `release-13-languages`；release-supported human UI locales 为 `en`、`zh-CN`、`ja`、`de`、`fr`、`es-ES`、`es-419`、`it`、`ko`、`pt-BR`、`ru`、`zh-TW`；README、runtime config、locale loaders、`src/i18n/catalog/messages/<lang>/{runtime,static,schema}.json`、`src/i18n/generated/locales/*.generated.ts` 与 WebExtension `_locales` 必须与该范围一致；当前 catalog-owned checked-in static source 为 `public/_locales/**`
 - `qps-ploc` 当前分类为 `dev-test-only` pseudo-locale：仅用于开发/测试伪本地化；production runtime locale registry、production build output、Chrome ZIP 与 Firefox XPI 均不得包含 `qps-ploc` loader/chunk 或 `_locales/qps-ploc/messages.json`
 - root `_locales/**` 已删除，不再作为 retained compatibility duplicate 保留；production build/package/release surface owner 为 `public/_locales/**` -> `build/dist/_locales/**` -> final Chrome ZIP / Firefox XPI
