+++ src/i18n/catalog/pseudoLocale.ts
@@ -1,12 +1,11 @@
-import type { LocaleDefinition, LocaleStaticMessages } from '../localeDefinition';
-import type { Messages } from '../messages';
+import type { LocaleDefinition, LocaleStaticMessages, RuntimeMessages } from '../localeDefinition';
 import en from '../generated/locales/en.generated';
 import { pseudoLocalizeMessages, pseudoLocalizeStatic } from '../pseudoLocalization';
 
 const ENGLISH_RUNTIME_MESSAGES = en.runtime;
 const ENGLISH_STATIC_MESSAGES = en.static;
 
-export function buildPseudoRuntimeMessages(): Messages {
+export function buildPseudoRuntimeMessages(): RuntimeMessages {
   return pseudoLocalizeMessages(ENGLISH_RUNTIME_MESSAGES);
 }
